+++ netlify/functions/user.ts
@@ -2,8 +2,6 @@ import type { Context } from "@netlify/functions";
 import type { User } from "../../src/types";
 import { fetchGraphQL } from "./utils";
 
-const GETUSERSURL = "https://shopping-cart.hasura.app/api/rest/getusers/";
-
 export default async (req: Request, context: Context) => {
 	const { method } = req;
 	const { params } = context;
@@ -11,19 +9,7 @@ export default async (req: Request, context: Context) => {
 	switch (method) {
 		case "GET":
 			try {
-				const resp: User[] = (
-					await (
-						await fetch(GETUSERSURL, {
-							headers: {
-								"Content-Type": "application/json",
-								"x-hasura-admin-secret":
-									process.env.VITE_HASURA_ADMIN_SECRET || "",
-							},
-						})
-					).json()
-				).user;
-
-				return new Response(JSON.stringify(resp || []), {
+				return new Response(JSON.stringify(await getAllUsers()), {
 					headers: {
 						"Content-Type": "application/json",
 					},
@@ -31,6 +17,7 @@ export default async (req: Request, context: Context) => {
 				});
 			} catch (error) {
 				console.error("Error fetching users:", error);
+
 				return new Response(
 					JSON.stringify({ error: "Failed to fetch users" }),
 					{
@@ -44,6 +31,13 @@ export default async (req: Request, context: Context) => {
 
 		case "POST": {
 			const username = decodeURI(params.username);
+
+			const allUsers = await getAllUsers();
+			const existingUser = allUsers.find((u) => u.username === username);
+			if (existingUser) {
+				return new Response(JSON.stringify(existingUser), { status: 201 });
+			}
+
 			const variables = {
 				username,
 			};
@@ -120,3 +114,16 @@ export default async (req: Request, context: Context) => {
 };
 
 export const config = { path: ["/api/user/:username", "/api/user"] };
+
+const GETUSERSURL = "https://shopping-cart.hasura.app/api/rest/getusers/";
+const getAllUsers = async (): Promise<User[]> =>
+	(
+		await (
+			await fetch(GETUSERSURL, {
+				headers: {
+					"Content-Type": "application/json",
+					"x-hasura-admin-secret": process.env.VITE_HASURA_ADMIN_SECRET || "",
+				},
+			})
+		).json()
+	).user || [];
