+++ src/components/index.ts
@@ -1,4 +1,3 @@
 export * from "./Admin";
 export * from "./Cart";
 export * from "./Inventory";
-export * from "./User";
